Add files via upload
Key Points About Observer Design Pattern :

Observer Patternfalls under behavioral design pattern.
An observer design pattern is very useful when you are interested in the state of an object and want to get notified whenever there is any change or update.
In an observer pattern, the object that keeps watch on the state of another object are called the object that is being watched is called Subject.
In this Pattern, we define a one-to-many relationship or dependency between objects so that when one object changes state than all its dependent objects are notified and updated automatically.
The subject has a list of observers to notify of any change in its state, so it provides a method using which observers can register and unregister themselves. The subject also contains a method to notify all the observers of any change or update and either it can send the update while notifying the observer or it can provide another method to get the update.